ml-model-train/API/requirements.txt
keras==2.15.0
tensorflow==2.15.0
joblib==1.4.2
torch==2.3.1
scikit-learn==1.3.2
opencv-python==4.10.0.84
pandas==2.2.2
numpy==1.26.4
#matplotlib
mediapipe==0.10.14
transformers==4.44.2
catboost==1.2.7

google-generativeai==0.7.2


torchaudio==2.3.1
librosa==0.10.2
nltk==3.8.1
datasets==2.20.0
soundfile==0.12.1
sentencepiece==0.2.0
sounddevice==0.5.0



python-multipart==0.0.6
python-dotenv==1.0.1
#uvicorn==0.30.5
uvicorn[standard]
fastapi== 0.112.0
pydantic==2.8.2

